refactor: create base context manager and validators
